Rename utils p2p_request to just request
